Services Process Document/requirements.txt
functions-framework==3.*
google-cloud-documentai==2.22.0
google-cloud-storage==2.14.0
google-cloud-aiplatform
google-cloud-bigquery
